type declarations moved to types.ts
