+++ frontend/src/components/JobsTable.test.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { describe, expect, it } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import JobsTable, { JobItem } from './JobsTable';
+import JobsTable from './JobsTable';
+import { JobItem } from '../types'
 
 describe('JobsTable', () => {
   const jobs: JobItem[] = [
